Divide File
