+++ Release-iOS/Release-iOS/Presentation/Activity/ActivityView.swift
@@ -74,53 +74,6 @@ struct ActivityView: View {
     }
 }
 
-private func activityCell(for activity: Activity) -> some View {
-    HStack {
-        Image(activity.image)
-            .resizable()
-            .frame(width: 96, height: 96)
-            .padding(.trailing, 16)
-        
-        VStack(alignment: .leading) {
-            HStack {
-                Text(activity.category)
-                    .font(.paragraph3)
-                    .foregroundColor(.gray5)
-                
-                Spacer()
-                
-                Text(activity.status)
-                    .font(.paragraph3)
-                    .foregroundColor(Color.black2)
-                    .padding(.horizontal, 17)
-                    .padding(.vertical, 3.5)
-                    .background(Color.primary1)
-                    .cornerRadius(8)
-            }
-            .padding(.bottom, 5)
-            
-            Text(activity.title)
-                .font(.heading4)
-                .foregroundColor(.gray1)
-            
-            Text(activity.content)
-                .font(.paragraph3)
-                .foregroundColor(.gray3)
-                .padding(.bottom, 8)
-            
-            Text(activity.person)
-                .font(.paragraph2)
-                .foregroundColor(.gray3)
-        }
-        Spacer()
-    }
-    .padding(.horizontal, 16)
-    .padding(.vertical, 12)
-    .background(Color(.black2))
-    .cornerRadius(16)
-    .listRowBackground(Color.clear)
-}
-
 struct ActivityView_Previews: PreviewProvider {
     static var previews: some View {
         ActivityView()
